Reorganize tests that require server running and add backup tests
