+++ .gitignore
@@ -1,7 +1,6 @@
 # Datalab database dirs - these can be removed once we have a better place for them
 files/
 uploads/
-server/
 ssh_config
 logs/
